feat(fe): 로그인 전/후 화면 통일화
